requirements.txt
annotated-types==0.7.0
anyio==4.7.0
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-speech==1.41.1
azure-common==1.1.28
azure-core==1.32.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
isodate==0.7.2
jiter==0.8.2
joblib==1.4.2
lxml==5.3.0
msrest==0.7.1
numpy==2.2.0
oauthlib==3.2.2
ollama==0.4.4
openai==1.58.1
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
